kali-swarm/requirements.txt
pyautogui==0.9.54
mss==9.0.1
requests==2.31.0
discord.py==2.3.2
python-dotenv==1.0.1
edge-tts==6.1.11
